requirements.txt
boto3==1.23.2
pandas==1.4.2
python-dotenv==0.20.0
selenium==4.2.0
SQLAlchemy==1.4.36
webdriver_manager==3.7.0
psycopg2==2.9.3
